requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.4
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
coverage==4.5.3
coveralls==1.7.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.1
django-cors-middleware==1.3.1
django-extensions==2.1.6
django-filter==2.1.0
django-simple-history==2.7.2
djangorestframework==3.9.2
docopt==0.6.2
entrypoints==0.3
extras==1.0.0
facebook-sdk==3.1.0
fixture==1.5.11
fixtures==3.0.0
flake8==3.7.7
future==0.17.1
google-auth==1.6.3
gunicorn==19.9.0
idna==2.8
linecache2==1.0.0
mccabe==0.6.1
more-itertools==7.0.0
oauthlib==3.0.1
pbr==5.2.0
pluggy==0.9.0
psycopg2==2.7.7
psycopg2-binary==2.8.2
py==1.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.7.1
pytest==4.4.1
pytest-cov==2.6.1
python-decouple==3.1
python-mimeparse==1.6.0
python-social-auth==0.3.6
python-twitter==3.5
python3-openid==3.1.0
pytz==2019.1
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
social-auth-core==3.1.0
sqlparse==0.3.0
testtools==2.3.0
traceback2==1.4.0
unittest2==1.1.0
urllib3==1.24.2
whitenoise==4.1.2
